Added early stop loss conditions to the train.py file
Added the train_clust file for training the GMM. There were underflow
issues with a model using float32, so I forced the model to turn into
float64, this greatly increased the amount of memory it took up. Up to
~22GB used during training.

Added functions to the visualisations.py file for visualising the
training loss of the AEC() Credit--> Jack.

Added functions from the original RISCluster github related to the
TSNE and other visualisations.

+++ data.py
@@ -99,7 +99,7 @@ class SpecgramShaper(object):
 class SpecgramToTensor(object):
     """Convert ndarrays in sample to Tensors."""
     def __call__(self, X):
-        return torch.from_numpy(X)
+        return torch.from_numpy(X).type(torch.DoubleTensor)
